refactor(Profile): separate components
